%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


% Other missing docs
\alias{as.character.RocCurve}
\alias{auc}
\alias{auc.RocCurve}
\alias{blockAvg}
\alias{findRocSmoothingForTpAtFp}
\alias{findRocSmoothingForTpAtFp.default}
\alias{getFpRate}
\alias{getFpRate.RocCurve}
\alias{getTpRate}
\alias{getTpRate.RocCurve}
\alias{lines.RocCurve}
\alias{plot.RocCurve}
\alias{points.RocCurve}
\alias{scanRocTpAtFp}
\alias{scanRocTpAtFp.default}
\alias{smooth}
\alias{cat}
\alias{require}
\alias{testSeparation}
\alias{parseRegion}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
  The rest are deprecated methods.
}

\author{Henrik Bengtsson}


\keyword{documentation}
\keyword{internal}
